finish refactor SocketHealthMonitor
