+++ packages/server/utils/relayUnsubscribe.ts
@@ -1,0 +1,10 @@
+import {ConnectedSubs} from 'socketHelpers/ConnectionContext'
+
+const relayUnsubscribe = (subs: ConnectedSubs | undefined, opId: string) => {
+  if (subs) {
+    subs[opId]?.return!()
+    delete subs[opId]
+  }
+}
+
+export default relayUnsubscribe
